requirements.txt
python-dotenv>=1.0.0
python-telegram-bot>=21.0.0
tzdata>=2024.1
pytest>=8.0.0

